// src/api/data/ContentItemType.ts
/* tslint:disable */
/* eslint-disable */
/*
 * Autogenerated by @creditkarma/thrift-typescript v3.7.6
 * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
*/
export enum ContentItemType {
    TEXT_FIELD = 0,
    COMBO_BOX = 1,
    USER_CHOICE = 2,
    CALENDAR = 3,
    TABLE = 4,
    SEPARATOR = 5,
    MULTILINE_TEXT_FIELD = 6,
    MARK = 7,
    CHECKBOX = 8,
    SWITCH_ITEM = 9,
    HTTP_LINK = 10,
    CALENDAR_RANGE = 11,
    CONTAINER = 12,
    HAND_BOOK = 13,
    ORG_STRUCTURE = 14,
    ATTACHMENT = 15,
    CURRENCY = 16,
    BUTTON = 17,
    DATE_TIME = 18,
    RICH_TEXT = 19,
    VOTE = 20,
    JIRA_TIME = 21
}